+++ test-run.sh
@@ -7,17 +7,16 @@ if [ $# -ne "1" ]; then
   exit 1
 fi
 
+echo "$(pwd)"
+
 ROOT_DIRRECTORY="$(pwd)"
 projectPath="$(cd "$(dirname "$1")"; pwd)/$(basename "$1")"
 graphMiner="DependencyGraph"
 graphAnalyzer="DependencyGraphAnalysis"
 BFCalculator="Truck-Factor"
 jetbrainsBFCalculator="risky-patterns-idea"
-outputs="outputs"
-# outputFolderName="$outputs/$(basename $projectPath)-$(date '+%Y-%m-%d-%H:%M')"
-outputFolderName="output"
+outputFolderName="output-test"
 rm -rf "$outputFolderName" 
-mkdir -p "$outputs"
 mkdir "$outputFolderName"
 
 outputFolderPath="$ROOT_DIRRECTORY/$outputFolderName"
@@ -32,14 +31,17 @@ avelinoBFResult="$outputFolderPath/avelinoBFResults.json"
 authorshipPath="$outputFolderPath/authorships.json"
 avelinoMergeFilePath="$ROOT_DIRRECTORY/$BFCalculator/gittruckfactor/repo_info/alias.txt"
 jetbrainsMergeFilePath="$projectPath/merged_emails.json"
+specialsFilePath="$projectPath/specials.txt"
 
-"./$graphMiner/gradlew" -p "./$graphMiner" extractDependencies -Pprojectpath="$projectPath" -Pgraphpath="$graphFilePath" -Ptargetdirectories="$targetDirectoriesPath"
+#Extract dependency graph
+cd "$ROOT_DIRRECTORY/$graphMiner"
+"./gradlew" extractDependencies -Pprojectpath="$projectPath" -Pgraphpath="$graphFilePath" -Ptargetdirectories="$targetDirectoriesPath" -Pspecials="$specialsFilePath"
 
-
+#Generate feature vector
 cd "$ROOT_DIRRECTORY/$graphAnalyzer"
-pip3 install -r requirements.txt
 python3 ./src/DependencyGraphEvaluator.py -g "$graphFilePath" -i "$graphImagePath" -f "$featuresFilePath"
 
+#Run avelino's tool
 cd "$ROOT_DIRRECTORY/$BFCalculator/gittruckfactor"
 rm -f "$avelinoMergeFilePath"
 touch "$avelinoMergeFilePath"
@@ -48,10 +50,18 @@ mvn package exec:java -Dexec.mainClass="aserg.gtf.GitTruckFactor" -Dexec.args="$
 
 # Jetbrains BF Calculation
 cd "$ROOT_DIRRECTORY/$jetbrainsBFCalculator"
-mv -f "$jetbrainsMergeFilePath" "$jetbrainsMergeFilePath-tmp"
+if [ -f "$jetbrainsMergeFilePath" ]
+then
+  mv -f "$jetbrainsMergeFilePath" "$jetbrainsMergeFilePath-tmp"
+fi
 "./gradlew" --stacktrace sigExport -Pprj="$projectPath" -Pout="$jetbrainsBFResult" -Ptarget="$targetDirectoriesPath" -Psig="$featuresFilePath" -Pauthorship="$authorshipPath"
-mv -f "$jetbrainsMergeFilePath-tmp" "$jetbrainsMergeFilePath"
 
+if [ -f "$jetbrainsMergeFilePath-tmp" ]
+then
+  mv -f "$jetbrainsMergeFilePath-tmp" "$jetbrainsMergeFilePath"
+fi
+
+#Generate resuts
 cd "$ROOT_DIRRECTORY/$graphAnalyzer"
-python3 ./src/OutputEvaluation.py -o "$resultsPath" -a "$authorshipPath" -i "$avelinoBFResult" "$jetbrainsBFResult"
+python3 ./src/OutputEvaluation.py -o "$resultsPath" -a "$authorshipPath" -s "$specialsFilePath" -i "$avelinoBFResult" "$jetbrainsBFResult"
